some ios7 compatability stuff
git-svn-id: https://subversion.assembla.com/svn/kinvey/ios-library/trunk@8024 868178e2-d38b-4ba8-a603-825f91c87007

+++ KinveyKit/KinveyKit/Source/KCSFileStore.m
@@ -161,7 +161,7 @@ static id lastRequest = nil;
     if (offset > 0) {
         NSUInteger remaining = length - offset;
         headers[@"Content-Length"] = [@(remaining) stringValue];
-        headers[@"Content-Range"] = [NSString stringWithFormat:@"bytes %llu-%d/%d",offset+1,length,length];
+        headers[@"Content-Range"] = [NSString stringWithFormat:@"bytes %llu-%lu/%lu",offset+1,(unsigned long)length,(unsigned long)length];
         [stream setProperty:@(offset) forKey:NSStreamFileCurrentOffsetKey];
     }
